nvme/{041-045,051,055}: double-quote def_* variable references
The following commit will add the code to set nvme command output to
def_* variables. This will trigger the ShellCheck warning SC2086. To
prepare for the change, double-quote the references to the def_*
variables. As for nvme/051 and 055, local variables are initialized with
def_subsysnqn, and they also require double quotes.

Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>