common/xfs: ignore first umount error on _xfs_mkfs_and_mount()
We want to help capture error messages with _xfs_mkfs_and_mount() on
$FULL, to do that we should avoid spamming error messages for things
which we know are not fatal. Such is the case of when we try to
mkfs a filesystem but before that try to umount the target path.
The first umount is just for sanity, so ignore the error messages from
it.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>
[Shin'ichiro: used --quiet option instead for redirect to dev/null]
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>