common: add and use min io for fio
When using fio we should not issue IOs smaller than the device supports.
Today a lot of places have in place 4k, but soon we will have devices
which support bs > ps. For those devices we should check the minimum
supported IO.

However, since we also have a min optimal IO, we might as well use that
as well. By using this we can also leverage the same lookup with stat
whether or not the target file is a block device or a file.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>
[Shin'ichiro: dropped minimum 4k size check in _fio_opts_to_min_io()]
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>