common/xfs: verify xfs_io supports statx atomic write attributes
xfs_io atomic write support is a dependency of the scsi and nvme atomic
write tests. The xfs_io atomic write support was introduced across
different versions so if xfs_io pwrite supports the -A (Atomic Write)
option, it doesn't necessarily support statx atomic write fields so that
needs to be verified separately.

Link: https://github.com/osandov/blktests/issues/161
Signed-off-by: Alan Adamson <alan.adamson@oracle.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>