nvme/063: fixup tls_key encryption check
The _nvme_ctrl_tls_key function returns 0 if `tls_key` exists in sysfs for
the respective nvme controller. This will be evaluated as true. However,
the test should error only if the key is not exposed by sysfs. Which
would mean the connection is not encrypted, as per the existing warning
message in the test. Currently, we are checking that it exists and
erroring out incorrectly. This patch fixes the above.

Link: https://github.com/osandov/blktests/issues/168
Fixes: 9aa2023312bf ("nvme: add testcase for secure concatenation")
Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>
Signed-off-by: Shin'ichiro Kawasaki <shinichiro.kawasaki@wdc.com>